# services/constructor/src/Models/RequestForQuotation.cs
/*
 * AwsomeConstructorCompany
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace AwsomeConstructor.Models
{ 
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public partial class RequestForQuotation : IEquatable<RequestForQuotation>
    { 
        /// <summary>
        /// Address of the property
        /// </summary>
        /// <value>Address of the property</value>

        [DataMember(Name="Address")]
        public string Address { get; set; }

        /// <summary>
        /// Start Date
        /// </summary>
        /// <value>Start Date</value>

        [DataMember(Name="StartDate")]
        public string StartDate { get; set; }

        /// <summary>
        /// Duration in days
        /// </summary>
        /// <value>Duration in days</value>

        [DataMember(Name="Duration")]
        public int? Duration { get; set; }

        /// <summary>
        /// Max budget in Euroes
        /// </summary>
        /// <value>Max budget in Euroes</value>

        [DataMember(Name="Budget")]
        public decimal? Budget { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class RequestForQuotation {\n");
            sb.Append("  Address: ").Append(Address).Append("\n");
            sb.Append("  StartDate: ").Append(StartDate).Append("\n");
            sb.Append("  Duration: ").Append(Duration).Append("\n");
            sb.Append("  Budget: ").Append(Budget).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (ReferenceEquals(null, obj)) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((RequestForQuotation)obj);
        }

        /// <summary>
        /// Returns true if RequestForQuotation instances are equal
        /// </summary>
        /// <param name="other">Instance of RequestForQuotation to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(RequestForQuotation other)
        {
            if (ReferenceEquals(null, other)) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    Address == other.Address ||
                    Address != null &&
                    Address.Equals(other.Address)
                ) && 
                (
                    StartDate == other.StartDate ||
                    StartDate != null &&
                    StartDate.Equals(other.StartDate)
                ) && 
                (
                    Duration == other.Duration ||
                    Duration != null &&
                    Duration.Equals(other.Duration)
                ) && 
                (
                    Budget == other.Budget ||
                    Budget != null &&
                    Budget.Equals(other.Budget)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    if (Address != null)
                    hashCode = hashCode * 59 + Address.GetHashCode();
                    if (StartDate != null)
                    hashCode = hashCode * 59 + StartDate.GetHashCode();
                    if (Duration != null)
                    hashCode = hashCode * 59 + Duration.GetHashCode();
                    if (Budget != null)
                    hashCode = hashCode * 59 + Budget.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(RequestForQuotation left, RequestForQuotation right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(RequestForQuotation left, RequestForQuotation right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}