# services/constructor/src/wwwroot/swagger-original.json
{
  "openapi" : "3.0.2",
  "info" : {
    "title" : "AwsomeConstructorCompany",
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "https://api.server.test/v1"
  } ],
  "tags" : [ {
    "name" : "RequestForQuotation",
    "description" : "Request for quoation"
  }, {
    "name" : "Project"
  } ],
  "paths" : {
    "/RequestForQuotation" : {
      "post" : {
        "tags" : [ "RequestForQuotation" ],
        "summary" : "Recives a new request for quoation from a HRC company and responds whether the request was submitted or not",
        "description" : "A HRC can sends a request for quoation via this API. AwsomeConstructor return whether the request was submitted or an exception happened.",
        "operationId" : "RequestForQuotationPOST",
        "requestBody" : {
          "description" : "RequestForQuotation",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RequestForQuotation"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Request submitted successfuly."
          },
          "default" : {
            "description" : "Exception happened!"
          }
        },
        "x-swagger-router-controller" : "Quotation"
      }
    },
    "/RequestForQuotation/{id}" : {
      "put" : {
        "tags" : [ "RequestForQuotation" ],
        "summary" : "Recives a request for editing the quotation request with specific id from a HRC company and responds whether the request was submitted or not",
        "description" : "A HRC can sends a request for editing the request for quoation with specific id  via this API. AwsomeConstructor return whether the request was submitted or an exception happened.",
        "operationId" : "RequestForQuotationPUT",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Id of the quotation request",
          "required" : true,
          "style" : "simple",
          "explode" : true,
          "schema" : {
            "type" : "integer"
          },
          "example" : 1
        } ],
        "requestBody" : {
          "description" : "RequestForQuotation",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RequestForQuotation"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Changes applied successfuly"
          },
          "404" : {
            "description" : "Resource not found"
          },
          "default" : {
            "description" : "Exception happened!"
          }
        },
        "x-swagger-router-controller" : "Quotation"
      },
      "delete" : {
        "tags" : [ "RequestForQuotation" ],
        "summary" : "Recives a request for deleting the quotation request with specific id from a HRC company and responds whether the quoation was deleted or exception happened.",
        "description" : "A HRC can sends a request for editing the request for quoation with specific id via this API. AwsomeConstructor return whether the request was deleted or exception happened.",
        "operationId" : "deleteRequestForQuotationIdDELETE",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Id of the quotation request",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          },
          "example" : 1
        } ],
        "responses" : {
          "200" : {
            "description" : "Deleted successfuly"
          },
          "404" : {
            "description" : "Resource not found"
          },
          "default" : {
            "description" : "Exception happened!"
          }
        },
        "x-swagger-router-controller" : "Quotation"
      }
    },
    "/Quotation/{id}" : {
      "get" : {
        "tags" : [ "RequestForQuotation" ],
        "summary" : "Recives a request for offering the quotation of the request for quotaion with specific id from a HRC company and responds with the Quotation.",
        "description" : "A HRC can sends a request for offering the Qutation of the request for quoation with specific id via this API. AwsomeConstructor responds with the Quotation.",
        "operationId" : "QuotationIdGET",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Id of the quotation request",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          },
          "example" : 1
        } ],
        "responses" : {
          "200" : {
            "description" : "Quotation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Quotation"
                }
              }
            }
          },
          "202" : {
            "description" : "Quotation not ready yet"
          },
          "default" : {
            "description" : "Exception happened!"
          }
        },
        "x-swagger-router-controller" : "Quotation"
      }
    },
    "/Project" : {
      "post" : {
        "tags" : [ "Project" ],
        "summary" : "Request to start the project",
        "description" : "In the event that a HRC decides AwsomeContructor as the winner, the HRC informs PlumberBrothersto start working on the project.",
        "operationId" : "ProjectPOST",
        "requestBody" : {
          "description" : "RequestForQuotation",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Project"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "ok"
          },
          "default" : {
            "description" : "Exception happened!"
          }
        },
        "x-swagger-router-controller" : "Project"
      }
    },
    "/Project/{id}" : {
      "get" : {
        "tags" : [ "Project" ],
        "summary" : "Gets completion status of the project",
        "description" : "The HRC company can check wether the project is completed or not.",
        "operationId" : "projectCompletionStatusIdGET",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Id of the quotation request",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          },
          "example" : 1
        } ],
        "responses" : {
          "200" : {
            "description" : "Ok",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "boolean",
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "default" : {
            "description" : "Exception happened!"
          }
        },
        "x-swagger-router-controller" : "Project"
      }
    }
  },
  "components" : {
    "schemas" : {
      "RequestForQuotation" : {
        "properties" : {
          "Address" : {
            "type" : "string",
            "description" : "Address of the property"
          },
          "StartDate" : {
            "type" : "string",
            "description" : "Start Date",
            "example" : "10/10/2023"
          },
          "Duration" : {
            "type" : "integer",
            "description" : "Duration in days",
            "example" : 60
          },
          "Budget" : {
            "type" : "number",
            "description" : "Max budget in Euroes",
            "example" : 1000
          }
        }
      },
      "Quotation" : {
        "properties" : {
          "id" : {
            "type" : "integer",
            "example" : 10
          },
          "servicePrice" : {
            "type" : "number",
            "description" : "the price for the service",
            "example" : 1000
          },
          "duration" : {
            "type" : "integer",
            "description" : "duration in days",
            "example" : 10
          },
          "startDate" : {
            "type" : "string",
            "description" : "the date that",
            "example" : "11/10/2023"
          }
        },
        "description" : "Model containg the quotation offered by the renovation company",
        "example" : {
          "duration" : 10,
          "servicePrice" : 1000,
          "id" : 10,
          "startDate" : "11/10/2023"
        }
      },
      "Project" : {
        "properties" : {
          "id" : {
            "type" : "integer",
            "example" : 10
          }
        }
      }
    }
  }
}