# services/eletrician/src/Controllers/ProjectApi.cs
/*
 * AwsomeElectricians
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using AwsomeElectricians.Attributes;

using Microsoft.AspNetCore.Authorization;
using AwsomeElectricians.Models;

namespace AwsomeElectricians.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class ProjectApiController : ControllerBase
    { 
        /// <summary>
        /// Request to start the project
        /// </summary>
        /// <remarks>In the event that a HRC decides AwsomeContructor as the winner, the HRC informs PlumberBrothersto start working on the project.</remarks>
        /// <param name="body">RequestForQuotation</param>
        /// <response code="200">ok</response>
        /// <response code="0">Exception happened!</response>
        [HttpPost]
        [Route("/v1/Project")]
        [ValidateModelState]
        [SwaggerOperation("NewProjectPOST")]
        public virtual IActionResult ProjectPOST([FromBody]Project body)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0);

            return Ok("Project Started Successfully");
        }
    }
}